Resource module
I have added the resource module with a basic documentation , cargo fmt doesn't seem to know what to do with the comments. I have also added the code for the sunray and energy cell taken from discord only to make sure it worked inside the resource module. this need to be fixed by the other contractor. This implementation of resource tries to fix the visibility problem of the combination and generation rules.